test(core): stop the project-graph recomputation specs hanging and leaking
Two problems made this file fail in CI on a different test each run, always
as a bare 35s timeout that named nothing.

`waitFor` had no deadline, and was copy-pasted eleven times. A condition that
never came true spun until vitest killed the test, so the report said only
that the test timed out. These cases run in 26-50ms, so a 35s failure was
always a hang rather than a slow test -- the timeout hid that. Each copy now
fails after ten seconds naming the condition it waited for.

The hang was cross-test contamination. The watcher-race cases stand up a real
file watcher, and its leftover events reach later tests as stray
recomputations; the warm-reuse case then sees an extra
retrieveProjectConfigurations and fails its exact-count assertion. Measured
under CPU saturation: that case passes 10/10 alone and its whole describe
8/8 alone, while the full file fails roughly one run in nine.

vitest only isolates per file (pool: 'forks'), so the watcher cases move to
their own spec and get their own process. Same load, same machine: 0/12
failures for the main file and 0/5 for the new one.
